Add ability to edit the records for artist and venue
The *.html files were fixed to have the abilities that
was visible in the forms but not the view.

A workaround for CSRF error that was being encountered was
incorporated as per the discussion on
https://knowledge.udacity.com/questions/67892. It is not clear if
the workaround is a proper way to do it. However keeping it for
the time being.