02_Trivia: Add delete and play quiz endpoints.
Added dataclass decorators to question model for easy jsonification
out of the box. SQL ORM objects have extra artifacts that make it
harder to jsonify, which complained of encountering non serializable
entities.

TODO: The frontend is broken which sends same codes for two categories.
Also there is a frontend issue in the page numbers not working for
the list tab. The former is critical, the latter lesser so.