02_Trivia: Allow all categories in quiz
Basically a very dirty hack to workaround the frontend not able
to send a consistent category ids. Tried to fix frontend, but it's
not clear where the id's are set. Trying to change the mapping
resulted in the first category vanishing altogether. The type
'click' is set **somewhere** if you pick all categories. Found it via
network inspection of the headers.