02_Trivia: Fix small bug and tests to be state independent.
Tests were hardcoded with values, but now are based on OQM queries,
which is maybe only slight improvement. But it's easier to test the
correct SQL queries are made for the given endpoint. Hardcoded value
tests would failk as adding and deleting can happen out of order and
there is no guarantee on the db state.

Resetting the entire db on teardown is too inefficient.